hedley.h: 2e88415 (Prefer deprecated GNU attribute over deprecated C++ attribute)
The GNU attribute has been supported for much longer, and compilers
which *claim* to support C++14 but don't completely support it (such as
xcode 7.3) do exist.  GCC and clang have traditionally added flags to
enable C and C++ standards and defined the relevant macros when they
start working on supporting a standard, not when they finish.

Fixes #41